Clean up rkstiff code

+++ src/rkstiff/ETDCS.cpp
@@ -15,20 +15,20 @@ ETD4::ETD4(const LinOp& Lop,const NLfunc& NL,bool use_refs) :
 void ETD4::updateCoefficients(double dt) noexcept
 {
     std::vector<dcmplx> r(contourPoints(),0.0);
-    for(int j = 0; j < contourPoints(); j++){
-        double expv = 2.0*PI*(j+0.5)/contourPoints();
+    for(unsigned j = 0; j < contourPoints(); j++){
+        double expv = 2.0*PI*(j+0.5)/static_cast<double>(contourPoints());
         r[j] = contourRadius()*exp(ii*expv);
     }
 
-    for(int i = 0; i < m_sz; i++) {
+    for(unsigned i = 0; i < m_sz; i++) {
         dcmplx Lval(dt*L[i]);
         EL[i] = exp(Lval);
         EL2[i] = exp(c2*Lval);
         dcmplx a21val(0.0,0.0); dcmplx a31val(0.0,0.0); dcmplx a32val(0.0,0.0);
-        dcmplx a41val(0.0,0.0); dcmplx a43val(0.0,0.0); 
+        dcmplx a41val(0.0,0.0); dcmplx a43val(0.0,0.0);
         dcmplx a51val(0.0,0.0); dcmplx a52val(0.0,0.0); dcmplx a54val(0.0,0.0);
         if(abs(Lval) < modeCutoff()){
-            for(int j = 0; j < contourPoints(); j++){
+            for(unsigned j = 0; j < contourPoints(); j++){
                 dcmplx z(Lval + r[j]);
                 a21val += (-1.0 + exp(z/2.0))/z;
                 a31val += (4.0 + exp(z/2.0)*(-4.0 + z) + z)/pow(z,2);
@@ -39,7 +39,7 @@ void ETD4::updateCoefficients(double dt) noexcept
                 a52val += (2.0*(2.0 + exp(z)*(-2.0 + z) + z))/pow(z,3);
                 a54val += -((4.0 + exp(z)*(-4.0 + z) + 3.0*z + pow(z,2))/pow(z,3));
             }
-            double mult = dt/contourPoints();
+            double mult = dt/static_cast<double>(contourPoints());
             a21[i] = mult*a21val; a31[i] = mult*a31val; a32[i] = mult*a32val;
             a41[i] = mult*a41val; a43[i] = mult*a43val;
             a51[i] = mult*a51val; a52[i] = mult*a52val; a54[i] = mult*a54val;
@@ -60,19 +60,19 @@ void ETD4::updateCoefficients(double dt) noexcept
 void ETD4::updateStages(std::vector<dcmplx>& in) noexcept
 {
     SolverCV::NL()(in,N1);
-    for(int i = 0; i < m_sz; i++)
+    for(unsigned i = 0; i < m_sz; i++)
         tempK[i] = EL2[i]*in[i] + a21[i]*N1[i];
-  
+
     SolverCV::NL()(tempK,N2);
-    for(int i = 0; i < m_sz; i++)
+    for(unsigned i = 0; i < m_sz; i++)
         tempK[i] = EL2[i]*in[i] + a31[i]*N1[i] + a32[i]*N2[i];
 
     SolverCV::NL()(tempK,N3);
-    for(int i = 0; i < m_sz; i++)
+    for(unsigned i = 0; i < m_sz; i++)
         tempK[i] = EL[i]*in[i] + a41[i]*N1[i] + a43[i]*N3[i];
 
     SolverCV::NL()(tempK,N4);
-    for(int i = 0; i < m_sz; i++) 
+    for(unsigned i = 0; i < m_sz; i++)
         in[i] = EL[i]*in[i] + a51[i]*N1[i] + a52[i]*(N2[i]+N3[i]) + a54[i]*N4[i];
 }
